requirement.txt
six==1.16.0
setuptools==58.2.0
ldap3==2.9
aniso8601==8.0.0
Click==7.0
DBUtils==1.3
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT==0.3.2
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
PyJWT==1.4.2
PyMySQL==0.9.3
xmltodict==0.12.0
gunicorn==20.1.0
pycryptodome==3.10.1
google-cloud-core==2.1.0
google-auth==2.3.0
google-cloud-storage==1.42.3
google-api-python-client==2.15.0
Flask-Docs==0.5.3
google-cloud-datacatalog==3.4.3
google-cloud-bigquery==2.28.1
flask-apscheduler==1.12.2